added seeders and did frontend

+++ backend/db/models/review.js
@@ -22,7 +22,7 @@ module.exports = (sequelize, DataTypes) => {
     spotId: DataTypes.INTEGER,
     userId: DataTypes.INTEGER,
     review: DataTypes.STRING,
-    stars: DataTypes.INTEGER
+    stars: DataTypes.DECIMAL
   }, {
     sequelize,
     modelName: 'Review',
